+++ src/test/java/me/alidg/errors/servlet/ServletApplication.java
@@ -1,0 +1,12 @@
+package me.alidg.errors.servlet;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class ServletApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(ServletApplication.class, args);
+    }
+}
