+++ src/test/resources/application.properties
@@ -1,0 +1,5 @@
+spring.messages.basename=test_messages
+errors.expose-arguments=always
+errors.add-fingerprint=true
+spring.main.banner-mode=off
+management.endpoints.web.exposure.include=metrics
